removes fft from Util, small code refactor
